+++ src/handlers/__init__.py
@@ -1,3 +1,3 @@
 from .command_handlers import (chat_info_handler, find_dolbaeb_handler,
-                               register_member_handler, start_handler)
+                               register_conversation_handler, start_handler)
 from .scheduled_handlers import bugin_ne_plan_handler
